add_cart/requirement.txt
sqlalchemy==1.0.11
redis==2.10.3
boto==2.38.0
lxml==3.4.4
psycopg2==2.6
Thrift==0.9.2
thriftpy==0.3.1
thrift_connector==0.7
jsonpickle==0.9.2
sortedcontainers==0.9.6
celery==3.1.18
requests==2.7.0
pytz==2015.7
redshift-sqlalchemy==0.4.1
simplejson==3.8.0
py-dateutil==2.2 
PyYAML==3.11
PySocks==1.5.6
beautifulsoup4==4.4.1
bottlenose==0.6.2
python-dateutil==2.4.2
python-amazon-simple-product-api==2.0.1
selenium==2.48.0
Pillow==3.0.0
PyVirtualDisplay==0.1.5
Jinja2==2.9.6
